combine find helpers into findElement and combine tests
- findText: prefer the element with the least descendants
